require PHP 7.3+ or PHP 8: ignore union types under PHP 8; upgrade dependencies; merge PHP 7.0 and 7.1 tests
