--- test/test-php70.php
@@ -1,14 +1,0 @@
-<?php
-
-use mindplay\unbox\Reflection;
-
-test(
-    'PHP 7.0: ignore scalar type-hints under PHP >= 7.0',
-    function () {
-        $reflection = new ReflectionFunction(function (string $foo) {});
-
-        $params = $reflection->getParameters();
-
-        eq(Reflection::getParameterType($params[0]), null);
-    }
-);
